Moved CreateMessageFragment to fragment package.| #1620
